t_server tofu: fix idempotency issue

+++ t_server/tofu/modules/tserver_user_data/provision/55-prepare_t_client.sh
@@ -9,7 +9,7 @@ cp -rv $OPENVPN_TESTS_GIT_REPO/t_server/original/client_vm/t_client.* $HOMEDIR/
 cp -rv $OPENVPN_TESTS_GIT_REPO/t_server/original/client_vm/bin/* $BINDIR/
 
 # Create a link to the "master" version of OpenVPN
-ln -s $BINDIR/openvpn $BINDIR/openvpn.master
+test -e $BINDIR/openvpn.master || ln -s $BINDIR/openvpn $BINDIR/openvpn.master
 
 # Modify t_client.rc files to match this environment. Variables come from
 # deployment-config.sh.
